Create <AccessibilityTestingSteps> component
